Completed and corrected task_1
