+++ src/main/java/JUnitHamcrestMockitoHomework/test/java/task_1/NegativeUnitTests/NegativeTestSuite.java
@@ -1,6 +1,7 @@
 package task_1.NegativeUnitTests;
 
 import org.junit.Test;
+import org.mockito.internal.matchers.Null;
 import task_1.PrimeNumbers;
 
 import java.util.ArrayList;
@@ -14,17 +15,17 @@ public class NegativeTestSuite {
     private PrimeNumbers prime = new PrimeNumbers();
 
     @Test
-    public void test_negative1(){
+    public void test_negative_below_zero(){
         Integer[] actual = prime.filterNumbers(-23);
         Integer[] expected = {};
         assertEquals("Массив содержит числа при отрицательных входных данных", expected, actual);
     }
 
     @Test
-    public void test_negative2(){
-        Integer[] actual = prime.filterNumbers(-100);
-        Integer[] expected = {};
-        assertEquals("Массив содержит числа при отрицательных входных данных", expected, actual);
+    public void test_negative_null(){
+        Integer[] actual = prime.filterNumbers(null);
+        Integer[] expected = null;
+        assertEquals("Массив не null при воодных данных null", expected, actual);
     }
 
 }
